refactor: 폴더 구조 변경 #17
파일을 동일한 이름의 폴더 내부로 옮김
