+++ frontend/src/common/components/Footer/Footer.tsx
