+++ frontend/src/common/components/MobileLayout/MobileLayout.tsx
